pyproject.toml
[tool.poetry]
name = "quantum-linear-systems"
version = "0.1.0"
description = "Quantum algorithms to solve linear systems of equations."
authors = ["dmark04 <dmark04@web.de>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9, <3.12" # >=3.9 because of amazon-braket, <3.12 because of classiq
classiq = "^0.31.0"
qiskit = ">=0.43.0, <0.45.0"
qiskit-algorithms = ">=0.2.1, <0.3.0"
matplotlib = "^3.5.0" # this version could possible be even lower
numpy = ">=1.23.0"
trackhhl = {git = "https://github.com/dmark04/TrackHHL"}
quantum-linear-solvers = "^0.1.1"
vqls-prototype = "^0.2.0"

[tool.poetry.group.aws.dependencies]
qiskit-braket-provider = "^0.0.5"
amazon-braket-sdk = "^1.61.0.post0"

[tool.poetry.group.test.dependencies]
pre-commit = "^3.5.0"
pytest-cov = "^4.1.0"
pytest = "^7.4.2"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "requires_auth: marks tests as requiring a classiq authentication token, thus being avoided on the pipeline",
]

[tool.poetry.scripts]
tests = "tests.run_tests:main"
show-cov = "tests.run_tests:open_coverage_report"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
